Switch test runner to SWC
