+++ langchain/src/agents/toolkits/index.ts
@@ -5,7 +5,7 @@ export {
   createOpenApiAgent,
 } from "./openapi/openapi.js";
 export {
-  VectorStoreInfo,
+  type VectorStoreInfo,
   VectorStoreToolkit,
   VectorStoreRouterToolkit,
   createVectorStoreAgent,
@@ -15,6 +15,6 @@ export { ZapierToolKit } from "./zapier/zapier.js";
 export { createRetrieverTool } from "./conversational_retrieval/tool.js";
 export {
   createConversationalRetrievalAgent,
-  ConversationalRetrievalAgentOptions,
+  type ConversationalRetrievalAgentOptions,
 } from "./conversational_retrieval/openai_functions.js";
 export { OpenAIAgentTokenBufferMemory } from "./conversational_retrieval/token_buffer_memory.js";
