+++ langchain/src/experimental/autogpt/index.ts
@@ -1,7 +1,7 @@
-export { AutoGPTPrompt, AutoGPTPromptInput } from "./prompt.js";
+export { AutoGPTPrompt, type AutoGPTPromptInput } from "./prompt.js";
 
 export { AutoGPTOutputParser, preprocessJsonInput } from "./output_parser.js";
 
-export { AutoGPT, AutoGPTInput } from "./agent.js";
+export { AutoGPT, type AutoGPTInput } from "./agent.js";
 
-export { AutoGPTAction } from "./schema.js";
+export type { AutoGPTAction } from "./schema.js";
