+++ langchain/src/experimental/babyagi/index.ts
@@ -1,4 +1,4 @@
 export { TaskCreationChain } from "./task_creation.js";
 export { TaskExecutionChain } from "./task_execution.js";
 export { TaskPrioritizationChain } from "./task_prioritization.js";
-export { BabyAGI, Task, BabyAGIInputs } from "./agent.js";
+export { BabyAGI, type Task, type BabyAGIInputs } from "./agent.js";
