+++ langchain/src/experimental/plan_and_execute/index.ts
@@ -3,10 +3,10 @@ export {
   BasePlanner,
   BaseStepContainer,
   BaseStepExecutor,
-  StepAction,
-  StepResult,
-  Step,
-  Plan,
+  type StepAction,
+  type StepResult,
+  type Step,
+  type Plan,
   ListStepContainer,
   LLMPlanner,
   ChainStepExecutor,
