+++ langchain/src/schema/tests/lib.ts
@@ -1,0 +1,143 @@
+/* eslint-disable no-promise-executor-return */
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+import { Runnable } from "../runnable/base.js";
+import {
+  AIMessage,
+  BaseMessage,
+  ChatResult,
+  GenerationChunk,
+} from "../index.js";
+import { BaseCallbackConfig } from "../../callbacks/manager.js";
+import { LLM } from "../../llms/base.js";
+import { BaseRetriever } from "../retriever.js";
+import { BaseChatModel } from "../../chat_models/base.js";
+import { BaseOutputParser } from "../output_parser.js";
+import { Document } from "../../document.js";
+
+/**
+ * Parser for comma-separated values. It splits the input text by commas
+ * and trims the resulting values.
+ */
+export class FakeSplitIntoListParser extends BaseOutputParser<string[]> {
+  lc_namespace = ["tests", "fake"];
+
+  getFormatInstructions() {
+    return "";
+  }
+
+  async parse(text: string): Promise<string[]> {
+    return text.split(",").map((value) => value.trim());
+  }
+}
+
+export class FakeRunnable extends Runnable<string, Record<string, any>> {
+  lc_namespace = ["tests", "fake"];
+
+  returnOptions?: boolean;
+
+  constructor(fields: { returnOptions?: boolean }) {
+    super(fields);
+    this.returnOptions = fields.returnOptions;
+  }
+
+  async invoke(
+    input: string,
+    options?: Partial<BaseCallbackConfig>
+  ): Promise<Record<string, any>> {
+    if (this.returnOptions) {
+      return options ?? {};
+    }
+    return { input };
+  }
+}
+
+export class FakeLLM extends LLM {
+  response?: string;
+
+  thrownErrorString?: string;
+
+  constructor(fields: { response?: string; thrownErrorString?: string }) {
+    super({});
+    this.response = fields.response;
+    this.thrownErrorString = fields.thrownErrorString;
+  }
+
+  _llmType() {
+    return "fake";
+  }
+
+  async _call(prompt: string): Promise<string> {
+    if (this.thrownErrorString) {
+      throw new Error(this.thrownErrorString);
+    }
+    return this.response ?? prompt;
+  }
+}
+
+export class FakeStreamingLLM extends LLM {
+  _llmType() {
+    return "fake";
+  }
+
+  async _call(prompt: string): Promise<string> {
+    return prompt;
+  }
+
+  async *_streamResponseChunks(input: string) {
+    for (const c of input) {
+      await new Promise((resolve) => setTimeout(resolve, 50));
+      yield { text: c, generationInfo: {} } as GenerationChunk;
+    }
+  }
+}
+
+export class FakeChatModel extends BaseChatModel {
+  _combineLLMOutput() {
+    return [];
+  }
+
+  _llmType(): string {
+    return "fake";
+  }
+
+  async _generate(
+    messages: BaseMessage[],
+    options?: this["ParsedCallOptions"]
+  ): Promise<ChatResult> {
+    if (options?.stop?.length) {
+      return {
+        generations: [
+          {
+            message: new AIMessage(options.stop[0]),
+            text: options.stop[0],
+          },
+        ],
+      };
+    }
+    const text = messages.map((m) => m.content).join("\n");
+    return {
+      generations: [
+        {
+          message: new AIMessage(text),
+          text,
+        },
+      ],
+      llmOutput: {},
+    };
+  }
+}
+
+export class FakeRetriever extends BaseRetriever {
+  lc_namespace = ["test", "fake"];
+
+  async _getRelevantDocuments(
+    _query: string
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  ): Promise<Document<Record<string, any>>[]> {
+    return [
+      new Document({ pageContent: "foo" }),
+      new Document({ pageContent: "bar" }),
+    ];
+  }
+}
